+++ src/services/dialogService.ts
@@ -9,6 +9,7 @@ import PromptDialogContent from '@/components/dialog/content/PromptDialogContent
 import SettingDialogContent from '@/components/dialog/content/SettingDialogContent.vue'
 import SignInContent from '@/components/dialog/content/SignInContent.vue'
 import TopUpCreditsDialogContent from '@/components/dialog/content/TopUpCreditsDialogContent.vue'
+import UpdatePasswordContent from '@/components/dialog/content/UpdatePasswordContent.vue'
 import ManagerDialogContent from '@/components/dialog/content/manager/ManagerDialogContent.vue'
 import ManagerHeader from '@/components/dialog/content/manager/ManagerHeader.vue'
 import ManagerProgressFooter from '@/components/dialog/footer/ManagerProgressFooter.vue'
@@ -363,6 +364,21 @@ export const useDialogService = () => {
     })
   }
 
+  /**
+   * Shows a dialog for updating the current user's password.
+   */
+  function showUpdatePasswordDialog() {
+    return dialogStore.showDialog({
+      key: 'global-update-password',
+      component: UpdatePasswordContent,
+      headerComponent: ComfyOrgHeader,
+      props: {
+        onSuccess: () =>
+          dialogStore.closeDialog({ key: 'global-update-password' })
+      }
+    })
+  }
+
   return {
     showLoadWorkflowWarning,
     showMissingModelsWarning,
@@ -377,6 +393,7 @@ export const useDialogService = () => {
     showApiNodesSignInDialog,
     showSignInDialog,
     showTopUpCreditsDialog,
+    showUpdatePasswordDialog,
     prompt,
     confirm
   }
